refactor: migrate Arc D v2 script logic to src/bid_euchre/arc_d_v2/ package
Extracts domain logic from 6 scripts in scripts/internal/ into the
src/bid_euchre/arc_d_v2/ package. Scripts become thin CLI wrappers
(~20-50 lines of argparse + main).

Migrations:
- rung_state.py RunState merged into schemas.py (single source of truth)
- run_rung.py -> orchestration.py (step execution, DAG, multi-seed, rerun)
- generate_advance_check.py -> advance_check.py (hypothesis eval, checks)
- generate_rung_tables.py -> tables.py (11 canonical CSV generators)
- generate_evidence_manifest.py -> manifest.py (provenance + markdown)
- generate_rung_report.py -> report.py (CSV-to-markdown rendering)

All 6 test files updated to import from package instead of importlib.util
script loading. Pure refactor -- no behavior changes. 166 tests pass.
